ow_mcp/pyproject.toml
[project]
name = "ow-mcp"
version = "0.1.0"
description = "Remote MCP service for Open Wearables"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.4",
    "cryptography>=45.0.4",
    "fastapi>=0.120.4",
    "fastmcp>=2.14.2",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
code-quality = [
    "ruff>=0.14.14",
]

[project.scripts]
start = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
